+++ test/libsolidity/syntaxTests/inheritance/override/common_base_and_unique_mention.sol
@@ -3,10 +3,10 @@ interface I {
     function g() external;
 }
 abstract contract A is I {
-    function f() external override {}
+    function f() external {}
 }
 abstract contract B is I {
-    function g() external override {}
+    function g() external {}
 }
 contract C is A, B {
 }