+++ test/libsolidity/syntaxTests/inheritance/override/diamond_interface_intermediate_public_state_variable_and_function_implemented.sol
@@ -3,11 +3,11 @@ interface I {
 }
 contract A is I
 {
-	uint public override f;
+	uint public f;
 }
 abstract contract B is I
 {
-	function f() external virtual override returns (uint) { return 2; }
+	function f() external virtual returns (uint) { return 2; }
 }
 abstract contract C is A, B {}
 // ----