+++ test/libsolidity/syntaxTests/inheritance/override/interface_exception/abstract_needed.sol
@@ -1,0 +1,7 @@
+abstract contract I {
+	function f() external virtual;
+}
+contract C is I {
+	function f() external {}
+}
+// ----