+++ test/libsolidity/syntaxTests/inheritance/override/interface_exception/diamond_needed.sol
@@ -1,0 +1,10 @@
+interface I {
+	function f() external;
+}
+interface J {
+	function f() external;
+}
+contract C is I, J {
+	function f() external {}
+}
+// ----