+++ test/libsolidity/syntaxTests/inheritance/override/override_interface.sol
@@ -1,9 +1,11 @@
 interface A {
 	function test() external returns (uint256);
 	function test2() external returns (uint256);
+	function test3() external returns (uint256);
 }
 contract X is A {
 	function test() external override returns (uint256) {}
 	function test2() external override(A) returns (uint256) {}
+	function test3() external returns (uint256) {}
 }
 // ----